+++ modules/core/index.ts
@@ -6,8 +6,7 @@ export * from "./src/dynamic-form.service";
 export * from "./src/dynamic-form-control.component";
 
 export * from "./src/checkbox/dynamic-checkbox.model";
-export * from "./src/input/dynamic-number-input.model";
-export * from "./src/input/dynamic-text-input.model";
+export * from "./src/input/dynamic-input.model";
 export * from "./src/radio/dynamic-radio.model";
 export * from "./src/select/dynamic-select.model";
 export * from "./src/textarea/dynamic-textarea.model";