+++ modules/core/src/dynamic-input-control.model.ts
@@ -2,6 +2,9 @@ import {DynamicFormControlModel} from "./dynamic-form-control.model";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_INPUT = "INPUT";
 
+export const DYNAMIC_FORM_INPUT_AUTOCOMPLETE_OFF = "off";
+export const DYNAMIC_FORM_INPUT_AUTOCOMPLETE_ON = "on";
+
 export abstract class DynamicInputControlModel<T> extends DynamicFormControlModel<T> {
 
     autoComplete: boolean;
@@ -17,8 +20,8 @@ export abstract class DynamicInputControlModel<T> extends DynamicFormControlMode
 
         super(configObject);
 
-        this.autoComplete = configObject["autoComplete"] === undefined ? true : configObject["autoComplete"];
-        this.autoFocus = configObject["autoFocus"] === undefined ? true : configObject["autoFocus"];
+        this.autoComplete = configObject["autoComplete"] === undefined ? DYNAMIC_FORM_INPUT_AUTOCOMPLETE_OFF : configObject["autoComplete"];
+        this.autoFocus = configObject["autoFocus"] === undefined ? false : configObject["autoFocus"];
         this.maxLength = configObject["maxLength"] || 100;
         this.placeholder = configObject["placeholder"] || "";
         this.prefix = configObject["prefix"] || null;