+++ modules/core/src/input/dynamic-input.model.spec.ts
@@ -1,14 +1,14 @@
-import {DYNAMIC_FORM_CONTROL_INPUT_TYPE_TEXT, DynamicTextInputModel} from "./dynamic-text-input.model";
-import {DYNAMIC_FORM_CONTROL_TYPE_INPUT} from "../dynamic-input-control.model";
+import {DYNAMIC_FORM_CONTROL_INPUT_TYPE_TEXT, DynamicInputModel} from "./dynamic-input.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_INPUT, DYNAMIC_FORM_INPUT_AUTOCOMPLETE_OFF} from "../dynamic-input-control.model";
 
-describe("DynamicTextInputModel test suite", () => {
+describe("DynamicInputModel test suite", () => {
 
     describe("default object test suite", () => {
 
-        let defaultObject: DynamicTextInputModel;
+        let defaultObject: DynamicInputModel;
 
         beforeEach(() => {
-            defaultObject = new DynamicTextInputModel({});
+            defaultObject = new DynamicInputModel({});
         });
         
         it("tests if correct default type property is set", () => {
@@ -21,6 +21,12 @@ describe("DynamicTextInputModel test suite", () => {
 
             expect(defaultObject.inputType).toBeDefined();
             expect(defaultObject.inputType).toEqual(DYNAMIC_FORM_CONTROL_INPUT_TYPE_TEXT);
+        });
+
+        it("tests if correct default autoComplete property is set", () => {
+
+            expect(defaultObject.autoComplete).toBeDefined();
+            expect(defaultObject.autoComplete).toEqual(DYNAMIC_FORM_INPUT_AUTOCOMPLETE_OFF);
         });
 
         it("tests if correct default prefix property is set", () => {