+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-# Changelog
+# alpha.4
+
+### **Breaking Changes**
+
+* DynamicTextInputModel and DynamicNumberInputModel condensed to DynamicInputModel
+* max, min, step attributes now working correctly